uccp420wlan: Fixes and Feature additions
Release 6.9 Updates.

1. Advertise support for cloned  skbs to avoid memcpy for TCP data
   transmissions
2. Move the FW Loading to insmod from ifconfig wlan0 up (PR#11)
3. Enable RTS/CTS protection from Minstrel (needed a kernel bug fix).
4. Increase the QUEUE flush timeout to handle 2 VIF cases.
5. Debug - add separate section in mac_stats for RPU frame Drops
6. Re-organized the HAL header files in to hal.h, rpu.h, soc.h so that
   it is easier for SoC desginer/customer to understand their scope.
   Also moved all common things to hal.h from hal_hostport.h
7. Bumped version to 6.0.6
